+++ app/src/main/java/com/edu/kotlindbnetwork/data/network/APIService.kt
@@ -1,6 +1,7 @@
 package com.edu.kotlindbnetwork.data.network
 
 import com.edu.kotlindbnetwork.Consts
+import com.edu.kotlindbnetwork.data.network.response.UserResponse
 import retrofit2.http.GET
 import retrofit2.http.Query
