+++ app/src/main/java/com/edu/kotlindbnetwork/repo/UserRepoNetwork.kt
@@ -3,7 +3,7 @@ package com.edu.kotlindbnetwork.repo
 import com.edu.kotlindbnetwork.data.network.APIService
 import com.edu.kotlindbnetwork.Consts
 import com.edu.kotlindbnetwork.data.db.user.User
-import com.edu.kotlindbnetwork.data.network.toModel
+import com.edu.kotlindbnetwork.data.network.response.toModel
 
 class UserRepoNetwork(
     private val api: APIService
